[structure] moved demo files, updated version
