+++ test_demo.py
@@ -1,4 +1,4 @@
-from dbi import demo
+import demo
 
 class TestDemo():
     def test_run_demo(self):
